# Adds precompile statements for way faster startup times