--- a/.github/workflows/runtests.yml
+++ b/.github/workflows/runtests.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: ['1.6', '1.7', '1.8']
+        julia-version: ['^1.9.0-0']
         os: [ubuntu-latest]
         provider: ['mkl', 'fftw']
       fail-fast: false