--- a/.gitignore
+++ b/.gitignore
@@ -6,7 +6,6 @@ gitignore
 .ipynb_checkpoints
 .tags
 .vscode
-dat
 scratch*
 **Untitled*.ipynb
 nb/
@@ -31,3 +30,9 @@ dev
 docs/src-staging/index.md
 poetry.lock
 pyproject.toml
+startup
+benchmarks
+dat/*
+!dat/compute_default_camb_Cls.jl
+!dat/default_camb_Cls.jld2
+*-workspace