--- a/dat/compute_default_camb_Cls.jl
+++ b/dat/compute_default_camb_Cls.jl
@@ -1,0 +1,3 @@
+using CMBLensing, JLD2, CodecZlib
+Cℓ = camb(ℓmax=16000);
+save(CMBLensing._default_Cℓs_path, "params", Cℓ.params, "Cℓ", Cℓ, compress=true)